query/pyproject.toml
[project]
name = "query"
version = "0.1.0"
description = "This project is a Test for habi company"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mysql-connector-python>=9.3.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[tool.mypy]
mypy_path = "src"
follow_imports = "normal"
explicit_package_bases = true